Student Information System
Student Information System
Student information System is to allow the administrator of any
organization to edit and find out the personal details of a student and
allows the student to keep up to date his profile .It’ll also facilitate
keeping all the records of students, such as their id, name, mailing
address, phone number etc. So all the information about an student will
be available in a few seconds.